Update README.md
This small project idea popped up in my head out of nowhere. it took me quite some time. But the end result was delicious. In this animation you will see alphabets popping from the left and right top sides of the screen and gathering together in about 5 seconds and making a keyboard type shape.